pyproject.toml
[tool.poetry]
name = "propli"
version = "0.0.1"
description = "Convert and view plist and yaml files"
authors = ["Trenten Oliver <trenteno@wepay.com>"]
readme = "README.md"
packages = [{ include = "propli", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
typer = { version = "^0.7.0", extras = ["all"] }
ruamel-yaml = "^0.17.21"
attrs = "^23.1.0"

[tool.poetry.scripts]
propli = "propli.cli:app"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
